refactor: share Codex plugin operations
